demrepofdave/al5int/fix_bugs_and_improve_code_quality_3: Native timestamps and routines put into it's own source and header files.  Added average runelk performance in ms to window title (updated every second)
