+++ src/host_abstraction_layer/allegro_4/video_internal.h
@@ -11,6 +11,22 @@
 #define uint16 unsigned short
 #define uint8 unsigned char
 
+/******************************************************************************
+* Typedefs
+*******************************************************************************/
+
+// This structure will all information related to the native window state
+// and coordinates.
+
+typedef struct
+{
+    int startx;           // Start position within the native window for the electron window.
+    int starty;           // Start position within the native window for the electron window.
+    int winsizex;         // Current window size in pixels
+    int winsizey;         // Current window size in pixels
+} window_info_elk_t;
+
+
 int  video_init_begin();
 void video_init_complete();
