+++ src/host_abstraction_layer/allegro_5/video_internal.h
@@ -1,16 +1,37 @@
 #ifndef _VIDEO_INTERNAL_H
 #define _VIDEO_INTERNAL_H
 
+/******************************************************************************
+* Include files
+*******************************************************************************/
+
 #include <allegro5/allegro.h>
 #include "host_abstraction_layer/video.h"
 
 
-// Private functions to be places in hal only header.
+/******************************************************************************
+* Preprocessor Macros
+*******************************************************************************/
 
 // 2xSaI routines.
 #define uint32 unsigned long
 #define uint16 unsigned short
 #define uint8 unsigned char
+
+/******************************************************************************
+* Typedefs
+*******************************************************************************/
+
+// This structure will all information related to the native window state
+// and coordinates.
+
+typedef struct
+{
+    int startx;           // Start position within the native window for the electron window.
+    int starty;           // Start position within the native window for the electron window.
+    int winsizex;         // Current window size in pixels
+    int winsizey;         // Current window size in pixels
+} window_info_elk_t;
 
 int  video_init_begin();
 void video_init_complete();
